fix(iris): decoupling sandbox from judge-handler (#2352)
* fix(iris): decopiling sandbox

* fix(iris): apply anonymous field

* fix(iris): remove unused files

* fix: use type inference

* fix(iris): restore delated changes
